+++ src/Operations/Map.cs
@@ -2,6 +2,7 @@ namespace Ametrin.Optional;
 
 partial struct Option
 {
+    [Obsolete("Use Match")]
     public TResult Map<TResult>(Func<TResult> success, Func<TResult> error)
         => _success ? success() : error();
 }
@@ -42,11 +43,13 @@ partial struct Result<TValue, TError>
 
 partial struct ErrorState
 {
+    [Obsolete("Use Match")]
     public TResult Map<TResult>(Func<TResult> success, Func<Exception, TResult> error) => _isError ? error(_error) : success();
 }
 
 partial struct ErrorState<TError>
 {
+    [Obsolete("Use Match")]
     public TResult Map<TResult>(Func<TResult> success, Func<TError, TResult> error) => _isError ? error(_error) : success();
 }